requirements.txt
meshtastic==2.7.0
Pypubsub==4.0.3
PyQt6==6.9.1
pyserial==3.5
urllib3==2.5.0
